build fixes vs2017 cuda 10

+++ alvr_server/ControlSocket.cpp
@@ -1,3 +1,4 @@
+
 #include <WS2tcpip.h>
 
 #include "ControlSocket.h"
@@ -104,7 +105,7 @@ bool ControlSocket::Recv(std::vector<std::string> &commands) {
 	buf[ret] = 0;
 	m_Buf += buf;
 
-	int index;
+	size_t index;
 	while ((index = m_Buf.find("\n")) != -1) {
 		commands.push_back(m_Buf.substr(0, index));
 		m_Buf.replace(0, index + 1, "");
@@ -134,6 +135,6 @@ void ControlSocket::SendCommandResponse(const char * commandResponse)
 {
 	if (m_ClientSocket != INVALID_SOCKET) {
 		// Send including NULL.
-		send(m_ClientSocket, commandResponse, strlen(commandResponse) + 1, 0);
+		send(m_ClientSocket, commandResponse, (int)strlen(commandResponse) + 1, 0);
 	}
 }
